2005-10-29  Stefan Jahn  <stefan@lkcc.org>
        * check_netlist.cpp: Skip range test if a property is assigned
        by a (sweep) variable.

        * precision.h: Improved --enable-double behaviour by
        introducing double representation dependent precision
        constants.

2005-10-27  Stefan Jahn  <stefan@lkcc.org>

        * parse_netlist.y: Fixed smaller memory leaks in the netlist
        parser.

        * vector.cpp, equation.cpp, evaluate.cpp, net.cpp: Fixed loads
        of memory leaks.

        * environment.cpp: Fixed memory leak (deleting substrates
        and constants).

        * dataset.cpp (load_touchstone), check_touchstone.cpp: Fixed
        scanner memory leak.

        * check_netlist.cpp (netlist_destroy): Calling
        netlist_lex_destroy() fixes a memory leak in the netlist
        scanner.


git-svn-id: https://qucs.svn.sourceforge.net/svnroot/qucs/trunk@689 b5b04e8c-4942-46c9-ab4f-83783d557d1c

+++ qucs-core/src/hbsolver.cpp
@@ -18,7 +18,7 @@
  * the Free Software Foundation, Inc., 51 Franklin Street - Fifth Floor,
  * Boston, MA 02110-1301, USA.  
  *
- * $Id: hbsolver.cpp,v 1.2 2005-06-02 18:17:50 raimi Exp $
+ * $Id: hbsolver.cpp,v 1.3 2005-10-31 16:15:31 ela Exp $
  *
  */
 
@@ -189,7 +189,7 @@ void hbsolver::collectHarmonics (void) {
 #if DEBUG
   fprintf (stderr, "%d frequencies: [ ", frequencies->getSize ());
   for (int i = 0; i < frequencies->getSize (); i++) {
-    fprintf (stderr, "%g ", real (frequencies->get (i)));
+    fprintf (stderr, "%g ", (double) real (frequencies->get (i)));
   }
   fprintf (stderr, "]\n");
 #endif /* DEBUG */
